use static linking to achieve virtual behavior of MCrypto::Backend
Backend definition depends of availability of OpenSSL and used options
but in general it can be determined statically. Therefore we dont need
special interface and additional cost of virtual table lookups. Instead we
separate part of implementation into 2 files and compile only one of them
depending on current project settings.